Move in-progress changes to the IN_DEVELOPMENT file
